Squashed commit of the following:
commit f662ae7a95878e64b601c19bcb60d6c72ccd491c
Author: Gustavo Vazquez <160083465+Great-Visions-Code@users.noreply.github.com>
Date:   Sat Mar 9 22:51:10 2024 -0800

    Update SettingsView.swift

commit 64b9069abe3f4d221a3df0dab7db726554969bff
Author: Gustavo Vazquez <160083465+Great-Visions-Code@users.noreply.github.com>
Date:   Sat Mar 9 22:51:06 2024 -0800

    Adjust font size to match settings

    Adjusted font size to match user settings

commit 8eb92f8345a42ba917728bdf05057575a49f4145
Author: Gustavo Vazquez <160083465+Great-Visions-Code@users.noreply.github.com>
Date:   Sat Mar 9 22:50:32 2024 -0800

    Update ChapterView.swift

    Adjusted fontSize to match user settings

commit ef89cd14cb5665759e12acc7c6168b2255bd3e13
Author: Gustavo Vazquez <160083465+Great-Visions-Code@users.noreply.github.com>
Date:   Sat Mar 9 22:50:13 2024 -0800

    Update AppColors.swift

commit d29e680b059eaae3a2282ef669859e5a15a5b73d
Author: Gustavo Vazquez <160083465+Great-Visions-Code@users.noreply.github.com>
Date:   Sat Mar 9 22:50:08 2024 -0800

    Update Survive.swift

commit 2eb1d9264e4ec45f55e03e6a258b80b26ce3fa5e
Author: Gustavo Vazquez <160083465+Great-Visions-Code@users.noreply.github.com>
Date:   Sat Mar 9 22:50:04 2024 -0800

    Update StoryDataModel.swift

commit c38d741130f12bd707f470927a7d3acf06076a2e
Author: Gustavo Vazquez <160083465+Great-Visions-Code@users.noreply.github.com>
Date:   Sat Mar 9 22:49:55 2024 -0800

    Update UserInterfaceState.xcuserstate

commit c92885b301a65226798d473526a0767ec5b5e8ad
Author: Gustavo Vazquez <160083465+Great-Visions-Code@users.noreply.github.com>
Date:   Sat Mar 9 20:41:54 2024 -0800

    Update UserInterfaceState.xcuserstate

commit 6ff857c7c2bf97c20c3245af10d728973003c463
Author: Gustavo Vazquez <160083465+Great-Visions-Code@users.noreply.github.com>
Date:   Sat Mar 9 20:41:52 2024 -0800

    Update WelcomeView.swift

    Hide navigation bar back button on WelcomeScreen()

commit f81532c995249d520203aada6a06dbf1d061ed96
Author: Gustavo Vazquez <160083465+Great-Visions-Code@users.noreply.github.com>
Date:   Fri Mar 8 22:53:04 2024 -0800

    Update UserInterfaceState.xcuserstate

commit 3c936db1aea244cd7be12d9fe59528c83c17c2cd
Author: Gustavo Vazquez <160083465+Great-Visions-Code@users.noreply.github.com>
Date:   Fri Mar 8 22:53:02 2024 -0800

    Update .gitignore

commit 99c6bf1dd0005c9f0cbfd419b1c14edebfdb12f1
Author: Gustavo Vazquez <160083465+Great-Visions-Code@users.noreply.github.com>
Date:   Fri Mar 8 22:51:44 2024 -0800

    Update WelcomeView.swift

    Changed NavigationView to NavigationStack

commit ab748a3f39849e488ed7162d0d4b7ba8b599b7b4
Author: Gustavo Vazquez <160083465+Great-Visions-Code@users.noreply.github.com>
Date:   Fri Mar 8 22:51:24 2024 -0800

    Update EndView.swift

commit 932f98f1f553ab36898351dd2798a4211c81985d
Author: Gustavo Vazquez <160083465+Great-Visions-Code@users.noreply.github.com>
Date:   Fri Mar 8 22:51:15 2024 -0800

    Update ChapterView.swift

    Removed unneeded var isLinkActive

commit 625b681b12a1116c5c8bf5eeeceb99417ff8a489
Author: Gustavo Vazquez <160083465+Great-Visions-Code@users.noreply.github.com>
Date:   Fri Mar 8 22:50:46 2024 -0800

    Update .gitignore

    Added *.xcuserstate

commit 0b50a4883c1b2fb7174b0b6cd27f38fda2e9cf15
Author: Gustavo Vazquez <160083465+Great-Visions-Code@users.noreply.github.com>
Date:   Wed Mar 6 16:48:04 2024 -0800

    Update UserInterfaceState.xcuserstate

+++ Survive/Views/NavigationButton.swift
@@ -52,7 +52,7 @@ struct NavigationButton<Destination: View>: View {
     var body: some View {
         NavigationLink(destination: destination) {
             Text(buttonText)
-                .font(.system(size: fontSize))
+                .font(.system(size: CGFloat(fontSize)))
                 .foregroundColor(fontColor)
                 .bold()
                 .padding()
@@ -78,5 +78,6 @@ struct ExampleUsageView: View {
 }
 
 #Preview {
-    NavigationButton(destination: WelcomeView(), buttonText: "Start")
+    NavigationButton(destination: WelcomeView(), 
+                     buttonText: "Start")
 }
